# src/main/src/main/resources/META-INF/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Apicurio Registry API [v2]",
        "version": "2.2.3-SNAPSHOT",
        "description": "Apicurio Registry is a datastore for standard event schemas and API designs. Apicurio Registry enables developers to manage and share the structure of their data using a REST interface. For example, client applications can dynamically push or pull the latest updates to or from the registry without needing to redeploy. Apicurio Registry also enables developers to create rules that govern how registry content can evolve over time. For example, this includes rules for content validation and version compatibility.\n\nThe Apicurio Registry REST API enables client applications to manage the artifacts in the registry. This API provides create, read, update, and delete operations for schema and API artifacts, rules, versions, and metadata. \n\nThe supported artifact types include:\n- Apache Avro schema\n- AsyncAPI specification\n- Google protocol buffers\n- GraphQL schema\n- JSON Schema\n- Kafka Connect schema\n- OpenAPI specification\n- Web Services Description Language\n- XML Schema Definition\n\n\n**Important**: The Apicurio Registry REST API is available from `https://MY-REGISTRY-URL/apis/registry/v2` by default. Therefore you must prefix all API operation paths with `../apis/registry/v2` in this case. For example: `../apis/registry/v2/ids/globalIds/{globalId}`.\n",
        "contact": {
            "name": "Apicurio",
...
        },
...
    },
    "paths": {
...
        "/admin/rules": {
            "summary": "Manage the global rules that apply to all artifacts if not otherwise configured.",
            "get": {
                "tags": [
                    "Global rules",
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RuleType"
                                    }
                                }
                            }
                        },
                        "description": "The list of names of the globally configured rules."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "listGlobalRules",
                "summary": "List global rules",
                "description": "Gets a list of all the currently configured global rules (if any).\n\nThis operation can fail for the following reasons:\n\n* A server error occurred (HTTP error `500`)\n"
            },
            "post": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Rule"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Global rules",
                    "Admin"
                ],
                "responses": {
                    "204": {
                        "description": "The global rule was added."
                    },
                    "400": {
                        "$ref": "#/components/responses/BadRequest"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "createGlobalRule",
                "summary": "Create global rule",
                "description": "Adds a rule to the list of globally configured rules.\n\nThis operation can fail for the following reasons:\n\n* The rule type is unknown (HTTP error `400`)\n* The rule already exists (HTTP error `409`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "delete": {
                "tags": [
                    "Global rules",
                    "Admin"
                ],
                "responses": {
                    "204": {
                        "description": "All global rules have been removed successfully."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "deleteAllGlobalRules",
                "summary": "Delete all global rules",
                "description": "Deletes all globally configured rules.\n\nThis operation can fail for the following reasons:\n\n* A server error occurred (HTTP error `500`)\n"
            }
        },
        "/admin/rules/{rule}": {
            "summary": "Manage the configuration of a single global artifact rule.",
            "get": {
                "tags": [
                    "Global rules",
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rule"
                                }
                            }
                        },
                        "description": "The global rule's configuration."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getGlobalRuleConfig",
                "summary": "Get global rule configuration",
                "description": "Returns information about the named globally configured rule.\n\nThis operation can fail for the following reasons:\n\n* Invalid rule name/type (HTTP error `400`)\n* No rule with name/type `rule` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Rule"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Global rules",
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Rule"
                                }
                            }
                        },
                        "description": "The global rule's configuration was successfully updated."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "updateGlobalRuleConfig",
                "summary": "Update global rule configuration",
                "description": "Updates the configuration for a globally configured rule.\n\nThis operation can fail for the following reasons:\n\n* Invalid rule name/type (HTTP error `400`)\n* No rule with name/type `rule` exists (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "delete": {
                "tags": [
                    "Global rules",
                    "Admin"
                ],
                "responses": {
                    "204": {
                        "description": "The global rule was successfully deleted."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "deleteGlobalRule",
                "summary": "Delete global rule",
                "description": "Deletes a single global rule.  If this is the only rule configured, this is the same\nas deleting **all** rules.\n\nThis operation can fail for the following reasons:\n\n* Invalid rule name/type (HTTP error `400`)\n* No rule with name/type `rule` exists (HTTP error `404`)\n* Rule cannot be deleted (HTTP error `409`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "parameters": [
                {
                    "name": "rule",
                    "description": "The unique name/type of a rule.",
                    "schema": {
                        "$ref": "#/components/schemas/RuleType"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
        "/admin/loggers": {
            "summary": "Manage logger settings/configurations.",
            "get": {
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NamedLogConfiguration"
                                    }
                                }
                            }
                        },
                        "description": "The list of logging configurations."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "listLogConfigurations",
                "summary": "List logging configurations",
                "description": "List all of the configured logging levels.  These override the default\nlogging configuration."
            }
        },
        "/admin/loggers/{logger}": {
            "summary": "Manage logger settings/configurations.",
            "get": {
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NamedLogConfiguration"
                                }
                            }
                        },
                        "description": "The logger configuration for the named logger."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getLogConfiguration",
                "summary": "Get a single logger configuration",
                "description": "Returns the configured logger configuration for the provided logger name. If no logger configuration is persisted, this returns the current default log configuration in the system."
            },
            "put": {
                "requestBody": {
                    "description": "The new logger configuration.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LogConfiguration"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NamedLogConfiguration"
                                }
                            }
                        },
                        "description": "The new configuration for the given logger."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "setLogConfiguration",
                "summary": "Set a logger's configuration",
                "description": "Configures the logger referenced by the provided logger name with the given configuration."
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NamedLogConfiguration"
                                }
                            }
                        },
                        "description": "The default logger configuration (now that the configuration for this logger has been removed, the \ndefault configuration is applied)."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "removeLogConfiguration",
                "summary": "Removes logger configuration",
                "description": "Removes the configured logger configuration (if any) for the given logger."
            },
            "parameters": [
                {
                    "name": "logger",
                    "description": "The name of a single logger.",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
        "/admin/export": {
            "summary": "Provides a way to export registry data.",
            "get": {
                "tags": [
                    "Admin"
                ],
                "parameters": [
                    {
                        "name": "forBrowser",
                        "description": "Indicates if the operation is done for a browser.  If true, the response will be a JSON payload with a property called `href`.  This `href` will be a single-use, naked download link suitable for use by a web browser to download the content.",
                        "schema": {
                            "type": "boolean"
                        },
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/zip": {
                                "schema": {
                                    "$ref": "#/components/schemas/File"
                                }
                            }
                        },
                        "description": "Response when the export is successful."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "exportData",
                "summary": "Export registry data",
                "description": "Exports registry data as a `.zip` archive."
            }
        },
        "/admin/import": {
            "summary": "Provides a way to import data into the registry.",
            "post": {
                "requestBody": {
                    "description": "The `.zip` file representing the previously exported registry data.",
                    "content": {
                        "application/zip": {
                            "schema": {
                                "$ref": "#/components/schemas/File"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Admin"
                ],
                "parameters": [
                    {
                        "name": "X-Registry-Preserve-GlobalId",
                        "description": "If this header is set to false, global IDs of imported data are ignored and replaced by the next ID in the global ID sequence. This allows importing any data even though the global IDs would cause a conflict.",
                        "schema": {
                            "type": "boolean"
                        },
                        "in": "header"
                    },
                    {
                        "name": "X-Registry-Preserve-ContentId",
                        "description": "If this header is set to false, content IDs of imported data are ignored and replaced by the next ID in the content ID sequence. The mapping between content and artifacts is preserved. This allows importing any data even though the content IDs would cause a conflict.",
                        "schema": {
                            "type": "boolean"
                        },
                        "in": "header",
                        "required": false
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Indicates that the import was successful."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "importData",
                "summary": "Import registry data",
                "description": "Imports registry data that was previously exported using the `/admin/export` operation."
            }
        },
...
        "/admin/config/properties": {
            "summary": "Manage configuration properties.",
            "get": {
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ConfigurationProperty"
                                    }
                                }
                            }
                        },
                        "description": "On a successful response, returns a list of configuration properties."
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "listConfigProperties",
                "summary": "List all configuration properties",
                "description": "Returns a list of all configuration properties that have been set.  The list is not paged.\n\nThis operation may fail for one of the following reasons:\n\n* A server error occurred (HTTP error `500`)\n"
            }
        },
        "/admin/config/properties/{propertyName}": {
            "summary": "Manage a single configuration property (by name).",
            "get": {
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ConfigurationProperty"
                                }
                            }
                        },
                        "description": "The configuration property value."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "getConfigProperty",
                "summary": "Get the value of a configuration property",
                "description": "Returns the value of a single configuration property.\n\nThis operation may fail for one of the following reasons:\n\n* Property not found or not configured (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "put": {
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateConfigurationProperty"
                            }
                        }
                    },
                    "required": true
                },
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "204": {
                        "description": "The configuration property was updated."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "updateConfigProperty",
                "summary": "Update a configuration property",
                "description": "Updates the value of a single configuration property.\n\nThis operation may fail for one of the following reasons:\n\n* Property not found or not configured (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "responses": {
                    "204": {
                        "description": "The configuration property was deleted."
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "operationId": "resetConfigProperty",
                "summary": "Reset a configuration property",
                "description": "Resets the value of a single configuration property.  This will return the property to\nits default value (see external documentation for supported properties and their default\nvalues).\n\nThis operation may fail for one of the following reasons:\n\n* Property not found or not configured (HTTP error `404`)\n* A server error occurred (HTTP error `500`)\n"
            },
            "parameters": [
                {
                    "name": "propertyName",
                    "description": "The name of a configuration property.",
                    "schema": {
                        "type": "string"
                    },
                    "in": "path",
                    "required": true
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
...
            "Rule": {
                "title": "Root Type for Rule",
                "description": "",
                "required": [
                    "config"
                ],
                "type": "object",
                "properties": {
                    "config": {
                        "type": "string"
                    },
                    "type": {
                        "$ref": "#/components/schemas/RuleType"
                    }
                },
                "example": {
                    "type": "VALIDITY",
                    "config": "FULL"
                }
            },
...
            "RuleType": {
                "description": "",
                "enum": [
                    "VALIDITY",
                    "COMPATIBILITY"
                ],
                "type": "string",
                "example": "VALIDITY",
                "x-codegen-package": "io.apicurio.registry.types"
            },
...
            "File": {
                "format": "binary",
                "type": "string",
                "x-codegen-inline": true
            }
        },
        "responses": {
...
            "ServerError": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "ErrorExample": {
                                "value": {
                                    "error_code": 500,
                                    "message": "Lost connection to the database."
                                }
                            }
                        }
                    }
                },
                "description": "Common response for all operations that can fail with an unexpected server error."
            },
            "BadRequest": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                },
                "description": "Common response for all operations that can return a `400` error."
            },
            "Conflict": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        },
                        "examples": {
                            "ConflictExample": {
                                "value": {
                                    "error_code": 409,
                                    "message": "The artifact content was invalid."
                                }
                            }
                        }
                    }
                },
                "description": "Common response used when an input conflicts with existing data."
            },
...
        }
    },
...
}